refactor(CLI): Unify argv to commands and options resolution
