+++ lib/Serverless.js
@@ -5,7 +5,6 @@ const BbPromise = require('bluebird');
 const os = require('os');
 const chalk = require('chalk');
 const updateNotifier = require('update-notifier');
-const minimist = require('minimist');
 const pkg = require('../package.json');
 const CLI = require('./classes/CLI');
 const Config = require('./classes/Config');
@@ -17,6 +16,7 @@ const Variables = require('./classes/Variables');
 const ServerlessError = require('./classes/Error').ServerlessError;
 const Version = require('./../package.json').version;
 const isStandaloneExecutable = require('./utils/isStandaloneExecutable');
+const resolveCliInput = require('./utils/resolveCliInput');
 
 const installationMaintananceCommands = new Set(['uninstall', 'upgrade']);
 
@@ -38,7 +38,8 @@ class Serverless {
     // use the servicePath from the options or try to find it in the CWD
     this.cliInputArgv = process.argv.slice(2);
     configObject.servicePath =
-      configObject.servicePath || this.utils.findServicePath(minimist(this.cliInputArgv).config);
+      configObject.servicePath ||
+      this.utils.findServicePath(resolveCliInput(this.cliInputArgv).options.config);
 
     this.config = new Config(this, configObject);
